% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignGroupToTaxa.R
\name{assignGroupToTaxa}
\alias{assignGroupToTaxa}
\title{Assigns each taxa variable with a group variable}
\usage{
assignGroupToTaxa(ListOfTaxa, AssignedGroup = NULL)
}
\arguments{
\item{ListOfTaxa}{A vector containing a list of all taxa in the analysis. All values are converted into Strings.}

\item{AssignedGroup}{A vector that matches each taxa to a group, or if not provided each taxa forms it's own group.}
}
\value{
A named vector matching each taxa to a group.
}
\description{
This function creates the grouping value for plotting, if no grouping variable is supplied, the taxa will
be grouped based on their own taxa name.
}
\examples{
taxaList <- c("Pinus", "Betula", "Poaceae", "Picea", "Artemisia")
groupList <- c("Trees", "Shrubs", "Grasses", "Trees", "Herbs")
assignGroupToTaxa(taxaList, groupList)

#>   Pinus    Betula   Poaceae     Picea      Artemisia
#>   "Trees"  "Shrubs" "Grasses"   "Trees"   "Herbs"

assignGroupToTaxa(taxaList)

#>  Pinus      Betula     Poaceae       Picea   Artemisia
#>  "Pinus"    "Betula"   "Poaceae"     "Picea" "Artemisia"
}
